requirements.txt
python-dotenv==1.0.0
numpy==1.24.3
pandas==1.5.3
tushare==1.2.89
baostock
yfinance==0.2.36
pandas-datareader==0.10.0
pytest
colorlog